feat: add initial domain entities, PostgreSQL config, and EF Core migration
- Add 10 core domain entities and 5 system enums
- Configure PostgreSQL AppDbContext and connection string
- Add initial EF Core migration 
- Added new enum and resolved enum mapping issues
- Implemented SaveChangesAsync timestamp interceptor for UpdatedAt properties

+++ src/ShopService/ShopService.Domain/Entities/OrderItem.cs
@@ -8,15 +8,15 @@ namespace ShopService.Domain.Entities
 {
     public class OrderItem
     {
-        public Guid Id { get; set; }
-        public Guid? OrderId { get; set; }
-        public Guid? ProductId { get; set; }
-        public string ProductName { get; set; } = string.Empty;
-        public int Quantity { get; set; }
-        public double UnitPrice { get; set; }
-        public double TotalPrice { get; set; }
+       public Guid Id { get; set; }
+public Guid? OrderId { get; set; }
+public Guid? ProductId { get; set; }
+public string ProductName { get; set; } = string.Empty;
+public int Quantity { get; set; }
+public decimal UnitPrice { get; set; }
+public decimal TotalPrice { get; set; }
 
-        public Order? Order { get; set; }
-        public Product? Product { get; set; }  
+public Order? Order { get; set; }
+public Product? Product { get; set; }  
     }
 }
