+++ src/ShopService/ShopService.Domain/Entities/Product.cs
@@ -9,26 +9,26 @@ namespace ShopService.Domain.Entities
     public class Product
     {
         public Guid Id { get; set; }
-        public string Name { get; set; } = string.Empty;
-        public string Slug { get; set; } = string.Empty;
-        public string Description { get; set; } = string.Empty;
-        public string? ShortDescription { get; set; }
-        public string Sku { get; set; } = string.Empty;
-        public decimal Price { get; set; }
-        public decimal? CompareAtPrice { get; set; }
-        public int StockQuantity { get; set; }
-        public int? LowStockThreshold { get; set; }
-        public bool IsActive { get; set; } = true;
-        public bool IsFeatured { get; set; }
-        public string? Brand { get; set; }
-        public double? Weight { get; set; }
-        public Guid? CategoryId { get; set; }
-        public DateTime CreatedAt { get; set; }
-        public DateTime? UpdatedAt { get; set; }
+public string Name { get; set; } = string.Empty;
+public string Slug { get; set; } = string.Empty;
+public string Description { get; set; } = string.Empty;
+public string? ShortDescription { get; set; }
+public string Sku { get; set; } = string.Empty;
+public decimal Price { get; set; }
+public decimal? CompareAtPrice { get; set; }
+public int StockQuantity { get; set; }
+public int? LowStockThreshold { get; set; }
+public bool IsActive { get; set; } = true;
+public bool IsFeatured { get; set; }
+public string? Brand { get; set; }
+public decimal? Weight { get; set; }
+public Guid? CategoryId { get; set; }
+public DateTime CreatedAt { get; set; }
+public DateTime? UpdatedAt { get; set; }
 
-        public Category? Category { get; set; }
-        public ICollection<OrderItem> OrderItems { get; set; } = new List<OrderItem>();
-        public ICollection<CartItem> CartItems { get; set; } = new List<CartItem>();
-        public ICollection<Review> Reviews { get; set; } = new List<Review>();
+public Category? Category { get; set; }
+public ICollection<OrderItem> OrderItems { get; set; } = new List<OrderItem>();
+public ICollection<CartItem> CartItems { get; set; } = new List<CartItem>();
+public ICollection<Review> Reviews { get; set; } = new List<Review>();
     }
 }
